Optimizations
Added some more properties to make the character class a little more of a template. Got rid of the moving property as it was useless. Scrapped the idea of using states instead of multiple variables as I need to check if multiple states are in effect to make animation logic decisions, perhaps in the future when I understand animation and canvas better I will be able to implement it.